ATTRIB: correction in member types of dxf_attrib_struct.

+++ src/attrib.h
@@ -2,7 +2,7 @@
  * \file attrib.h
  *
  * \author Copyright (C) 2008, 2010, 2012, 2013, 2014, 2015, 2016, 2017,
- * 2018 by Bert Timmerman <bert.timmerman@xs4all.nl>.
+ * 2018, 2019 by Bert Timmerman <bert.timmerman@xs4all.nl>.
  *
  * \brief Header file for a DXF attribute entity (\c ATTRIB).
  *
@@ -94,20 +94,20 @@ dxf_attrib_struct
                  * </ol>
                  * Group code = 60.\n
                  * \since Introduced in version R13. */
-        int color;
+        int16_t color;
                 /*!< Color of the entity.\n
                  * Defaults to \c BYLAYER if ommitted in the DXF file.\n
                  * Note that entities encapsulated in a block with the
                  * color \c BYBLOCK are represented in the "native" color of
                  * the \c BLOCK entity.\n
                  * Group code = 62. */
-        int paperspace;
+        int16_t paperspace;
                 /*!< Entities are to be drawn on either \c PAPERSPACE or
                  * \c MODELSPACE.\n
                  * Optional, defaults to \c DXF_MODELSPACE (0).\n
                  * Group code = 67.\n
                  * \since Introduced in version R13. */
-        int graphics_data_size;
+        int32_t graphics_data_size;
                 /*!< Number of bytes in the proxy entity graphics
                  * represented in the sub-sequent 310 groups, which are
                  * binary chunk records (optional).\n
@@ -219,7 +219,7 @@ dxf_attrib_struct
         double obl_angle;
                 /*!< Oblique angle (optional; default = 0.0).\n
                  * Group code = 51. */
-        int attr_flags;
+        int16_t attr_flags;
                 /*!< Attribute flags.\n
                  * Bit coded:\n
                  * <ol>
@@ -232,7 +232,7 @@ dxf_attrib_struct
                  * during insertion).</li>
                  * </ol>
                  * Group code = 70. */
-        int text_flags;
+        int16_t text_flags;
                 /*!< Text generation flags (optional, default = 0).\n
                  * See \c TEXT.\n
                  * Optional, defaults to 0\n
@@ -242,7 +242,7 @@ dxf_attrib_struct
                  * <li value = "4"> Text is upside down (mirrored in Y).</li>
                  * </ol>
                  * Group code = 71. */
-        int hor_align;
+        int16_t hor_align;
                 /*!< Horizontal text justification type (optional,
                  * default = 0).\n
                  * See \c TEXT.\n
@@ -256,11 +256,11 @@ dxf_attrib_struct
                  * <li value = "5"> Fit, only when vert_align = 0.</li>
                  * </ol>
                  * Group code = 72. */
-        int field_length;
+        int16_t field_length;
                 /*!< Field length (optional; default = 0) (not currently
                  * used in DXF R14).\n
                  * Group code = 73. */
-        int vert_align;
+        int16_t vert_align;
                 /*!< Vertical text justification type (optional,
                  * default = 0).\n
                  * See group code 73 in \c TEXT.\n
@@ -311,12 +311,12 @@ double dxf_attrib_get_linetype_scale (DxfAttrib *attrib);
 DxfAttrib *dxf_attrib_set_linetype_scale (DxfAttrib *attrib, double linetype_scale);
 int16_t dxf_attrib_get_visibility (DxfAttrib *attrib);
 DxfAttrib *dxf_attrib_set_visibility (DxfAttrib *attrib, int16_t visibility);
-int dxf_attrib_get_color (DxfAttrib *attrib);
-DxfAttrib *dxf_attrib_set_color (DxfAttrib *attrib, int color);
-int dxf_attrib_get_paperspace (DxfAttrib *attrib);
-DxfAttrib *dxf_attrib_set_paperspace (DxfAttrib *attrib, int paperspace);
-int dxf_attrib_get_graphics_data_size (DxfAttrib *attrib);
-DxfAttrib *dxf_attrib_set_graphics_data_size (DxfAttrib *attrib, int graphics_data_size);
+int16_t dxf_attrib_get_color (DxfAttrib *attrib);
+DxfAttrib *dxf_attrib_set_color (DxfAttrib *attrib, int16_t color);
+int16_t dxf_attrib_get_paperspace (DxfAttrib *attrib);
+DxfAttrib *dxf_attrib_set_paperspace (DxfAttrib *attrib, int16_t paperspace);
+int32_t dxf_attrib_get_graphics_data_size (DxfAttrib *attrib);
+DxfAttrib *dxf_attrib_set_graphics_data_size (DxfAttrib *attrib, int32_t graphics_data_size);
 int16_t dxf_attrib_get_shadow_mode (DxfAttrib *attrib);
 DxfAttrib *dxf_attrib_set_shadow_mode (DxfAttrib *attrib, int16_t shadow_mode);
 DxfBinaryGraphicsData *dxf_attrib_get_binary_graphics_data (DxfAttrib *attrib);
@@ -369,20 +369,20 @@ double dxf_attrib_get_rot_angle (DxfAttrib *attrib);
 DxfAttrib *dxf_attrib_set_rot_angle (DxfAttrib *attrib, double rot_angle);
 double dxf_attrib_get_obl_angle (DxfAttrib *attrib);
 DxfAttrib *dxf_attrib_set_obl_angle (DxfAttrib *attrib, double obl_angle);
-int dxf_attrib_get_attr_flags (DxfAttrib *attrib);
-DxfAttrib *dxf_attrib_set_attr_flags (DxfAttrib *attrib, int attr_flags);
+int16_t dxf_attrib_get_attr_flags (DxfAttrib *attrib);
+DxfAttrib *dxf_attrib_set_attr_flags (DxfAttrib *attrib, int16_t attr_flags);
 int dxf_attrib_is_invisible (DxfAttrib *attrib);
 int dxf_attrib_is_constant (DxfAttrib *attrib);
 int dxf_attrib_is_verification_required (DxfAttrib *attrib);
 int dxf_attrib_is_preset (DxfAttrib *attrib);
-int dxf_attrib_get_text_flags (DxfAttrib *attrib);
-DxfAttrib *dxf_attrib_set_text_flags (DxfAttrib *attrib, int text_flags);
-int dxf_attrib_get_hor_align (DxfAttrib *attrib);
-DxfAttrib *dxf_attrib_set_hor_align (DxfAttrib *attrib, int hor_align);
-int dxf_attrib_get_field_length (DxfAttrib *attrib);
-DxfAttrib *dxf_attrib_set_field_length (DxfAttrib *attrib, int field_length);
-int dxf_attrib_get_vert_align (DxfAttrib *attrib);
-DxfAttrib *dxf_attrib_set_vert_align (DxfAttrib *attrib, int vert_align);
+int16_t dxf_attrib_get_text_flags (DxfAttrib *attrib);
+DxfAttrib *dxf_attrib_set_text_flags (DxfAttrib *attrib, int16_t text_flags);
+int16_t dxf_attrib_get_hor_align (DxfAttrib *attrib);
+DxfAttrib *dxf_attrib_set_hor_align (DxfAttrib *attrib, int16_t hor_align);
+int16_t dxf_attrib_get_field_length (DxfAttrib *attrib);
+DxfAttrib *dxf_attrib_set_field_length (DxfAttrib *attrib, int16_t field_length);
+int16_t dxf_attrib_get_vert_align (DxfAttrib *attrib);
+DxfAttrib *dxf_attrib_set_vert_align (DxfAttrib *attrib, int16_t vert_align);
 double dxf_attrib_get_extr_x0 (DxfAttrib *attrib);
 DxfAttrib *dxf_attrib_set_extr_x0 (DxfAttrib *attrib, double extr_x0);
 double dxf_attrib_get_extr_y0 (DxfAttrib *attrib);
